+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/response/index.ts
@@ -1,0 +1,1 @@
+export { response } from './response.js'
