+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/response/response.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { response } from './response'
+import { response } from '.'
 
 // Test run
 // pnpm vitest run ./src/generator/zod-openapi-hono/openapi/route/response/generate-response-schema.test.ts
